% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenaccesstotal.R
\name{access.model.total}
\alias{access.model.total}
\title{Modelling of public access to green space and natural ecosystems and open water - Totals}
\usage{
access.model.total(lcm, noentry = NULL, pop)
}
\arguments{
\item{lcm}{Land/ water cover map, including vegetation categories. Categories must follow Gaw et al. 2019, described in data(looktbl). Must be a raster with map units in metres.}

\item{noentry}{Raster indicating areas of no entry with values of 1. Function can also be used without this value, defaults to no inaccessible space in the study area.}

\item{pop}{Population given as a SpatialPointsDataFrame with one field for population, which is named "pop"}
}
\value{
A stack of three raster maps indicating the population pressure (number of people) with access to each pixel of (1) accessible green space in a patch larger than 1 ha, (2) blue space, (3) natural space.
}
\description{
This function models public access to green space, natural ecosystems, and open water and gives the total population with access to each.
}
