% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtemperature.R
\name{air.temp.model}
\alias{air.temp.model}
\title{Modelling of reduction in air temperature caused by ecosystems}
\usage{
air.temp.model(lcm, lai)
}
\arguments{
\item{lcm}{Land/ water cover map, including vegetation categories. Categories must follow Gaw et al. 2019, described in data(looktbl)}

\item{lai}{Leaf area index - raster grid with cell sizes in metre units}
}
\value{
One raster indicating average reduction in temperature at 15:00
}
\description{
This function models reduction in air temperature caused by ecosystems, in comparison to urban land cover. It is based on a Singapore-specific land use regression so is not useful in other cities.
}
