% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lai.converter.R
\name{lai.converter}
\alias{lai.converter}
\title{Conversion of partial LAI to full LAI}
\usage{
lai.converter(lcm, lai, vegheight, instrumentheight)
}
\arguments{
\item{lcm}{Land/ water cover map, including vegetation categories. Categories must follow Gaw et al. 2019, described in data(looktbl)}

\item{lai}{Leaf area index - raster grid with cell sizes in metre units}

\item{vegheight}{Vegetation height in metres}

\item{instrumentheight}{Height in metres of instrument used to measure LAI}
}
\value{
One raster of full LAI
}
\description{
This function does a simple conversion to estimate full LAI from a map of LAI above a certain point
}
