% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{noise.model}
\alias{noise.model}
\title{Modelling of noise attenuation by vegetation}
\usage{
noise.model(
  lcm,
  dem,
  buildings,
  roads,
  dB = 70,
  erandom = 100,
  humidity = 70,
  temperature = 30,
  frequency = 1000
)
}
\arguments{
\item{lcm}{Land/ water cover map, including vegetation categories. Categories must follow Gaw et al. 2019, described in data(looktbl). Must be a raster with map units in metres.}

\item{dem}{Raster digital elevation model in same resolution as land cover map.}

\item{buildings}{PBuildings given as a SpatialPointsDataFrame with one field for building height.}

\item{roads}{Rasterised road network in same resolution as land cover map.}

\item{dB}{Decibels for road traffic noise sources.}

\item{erandom}{Number of random noise sources to generate (has direct impact on processing time)}

\item{humidity}{Humidity of the study area in %, defaults to Singapore typical conditions.}

\item{temperature}{Temperature of the study area in degrees celsius, defaults to Singapore typical conditions.}

\item{frequency}{Frequency of the road traffic noise for analysis. Defaults to typical road traffic noise conditions.}
}
\value{
A stack of three raster maps indicating the population pressure (number of people) with access to each pixel of (1) accessible green space in a patch larger than 1 ha, (2) blue space, (3) natural space.
}
\description{
This function models soil erosion
}
