% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runoff.R
\name{runoff.model}
\alias{runoff.model}
\title{Modelling of runoff absorbtion by vegetation}
\usage{
runoff.model(lcm, rainfallin)
}
\arguments{
\item{lcm}{Land/ water cover map, including vegetation categories. Categories must follow Gaw et al. 2019, described in data(looktbl)}

\item{rainfallin}{Rainfall in mm}
}
\value{
Two rasters in a list, the first is proportion of runoff absorbed. Second is runoff in mm
}
\description{
This function models runoff reduction by vegetation
}
