% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilK.R, R/soilerosion.R
\name{soilerosion.model}
\alias{soilerosion.model}
\title{Calculation of soil K factor based on soil data}
\usage{
soilerosion.model(lcm, R, K, LS, ndvi)

soilerosion.model(lcm, R, K, LS, ndvi)
}
\arguments{
\item{lcm}{Land/ water cover map, including vegetation categories. Categories must follow Gaw et al. 2019, described in data(looktbl). Must be a raster with map units in metres.}

\item{R}{Rainfall erosivity (MJ/  mm / year). Can be either a raster or single value}

\item{K}{Soil erodibility factor. Can be either a raster with K values for different land cover types or a single value}

\item{LS}{Topographic factors. Must be a raster. Creating an LS raster is straightforward but computationally expensive in R, so a function is not given.}

\item{ndvi}{Normalised difference vegetation index}

\item{OM}{Percentage organic matter}

\item{Cm}{percentage clay by mass}

\item{Dg}{Geometric mean of the particle size distribution}
}
\value{
K values

List of three rasters - the ecosystem service ES, structural impact SI, and absolute soil erosion A
}
\description{
This function estimates a soil K factor for a given soil condition

This function models soil erosion
}
