% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegnoise.R
\name{veg.noise.model}
\alias{veg.noise.model}
\title{Modelling of noise attenuation by vegetation - model for fast modelling of veg only}
\usage{
veg.noise.model(
  lcm2,
  dB = 70,
  erandom = 100,
  humidity = 70,
  temperature = 30,
  frequency = 1000,
  roadsX,
  ssl,
  aal,
  bar
)
}
\arguments{
\item{lcm2}{Land/ water cover map, including vegetation categories. Categories must follow Gaw et al. 2019, described in data(looktbl). Must be a raster with map units in metres.}

\item{dB}{Decibels for road traffic noise sources.}

\item{erandom}{Number of random noise sources to generate (has direct impact on processing time)}

\item{humidity}{Humidity of the study area in %, defaults to Singapore typical conditions.}

\item{temperature}{Temperature of the study area in degrees celsius, defaults to Singapore typical conditions.}

\item{frequency}{Frequency of the road traffic noise for analysis. Defaults to typical road traffic noise conditions.}

\item{roadsX}{Rasterised road network in same resolution as land cover map.}

\item{ssl}{SSL layer from output of noise.model}

\item{aal}{AAL layer from output of noise.model}

\item{bar}{BAR layer from output of noise.model}
}
\value{
A stack of 6 raster maps, representing (1) noise after Spherical spreading, (2) noise after Atmospheric absorption, (3) noise after Barrier loss, (4) noise after Vegetation loss, (5) the vegetation ecosystem service, and (6) the location of noise sources.
}
\description{
This function speeds up models of noise attenuation by vegetation cover, if only the veg cover is changing in the new scenario. It is used to update an existing modelled scenario with identical noise sources, buildings, topography, etc.
}
