% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm10.R
\name{pm10.model}
\alias{pm10.model}
\title{Modelling of PM10 removal by vegetation}
\usage{
pm10.model(lai, lcm, pbl, pm10)
}
\arguments{
\item{lai}{Leaf area index - raster grid with cell sizes in metre units}

\item{pbl}{Height of the planetary boundary layer in metres}

\item{pm10}{PM10 concentration in atmosphere (micrograms per m3)}
}
\value{
List with one raster of Daily  PM10 removal in micrograms. Second is the percentage decrease in PM10 concentration.
}
\description{
This function models PM10 removal by vegetation
}
