% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilK.R
\name{soilK.model}
\alias{soilK.model}
\title{Calculation of soil K factor based on soil data}
\usage{
soilK.model(OM, C, Dg)
}
\arguments{
\item{OM}{Percentage organic matter}

\item{Dg}{Geometric mean of the particle size distribution}

\item{Cm}{percentage clay by mass}
}
\value{
K values
}
\description{
This function estimates a soil K factor for a given soil condition
}
